+++ lib/mfynab/cli.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "yaml"
+require "mfynab/config"
 require "mfynab/money_forward"
 require "mfynab/money_forward/session"
 require "mfynab/money_forward_data"
@@ -18,20 +19,9 @@ module MFYNAB
 
     def start
       logger.info("Running...")
-
       money_forward.update_accounts(money_forward_account_names)
-
-      Dir.mktmpdir("mfynab") do |save_path|
-        money_forward.download_csv(
-          path: save_path,
-          months: months_to_sync,
-        )
-
-        data = MoneyForwardData.new(logger: logger)
-        data.read_all_csv(save_path)
-        ynab_transaction_importer.run(data.to_h)
-      end
-
+      data = money_forward.fetch_data(config.months_to_sync)
+      ynab_transaction_importer.run(data.to_h)
       logger.info("Done!")
     end
 
@@ -40,20 +30,16 @@ module MFYNAB
       attr_reader :argv
 
       def money_forward_account_names
-        @_money_forward_account_names = config["accounts"].map { _1["money_forward_name"] }
+        @_money_forward_account_names = config.accounts.map { _1["money_forward_name"] }
       end
 
       def ynab_transaction_importer
         @_ynab_transaction_importer ||= YnabTransactionImporter.new(
-          config["ynab_access_token"],
-          config["ynab_budget"],
-          config["accounts"],
+          config.credentials["ynab_access_token"],
+          config.ynab_budget,
+          config.accounts,
           logger: logger,
         )
-      end
-
-      def months_to_sync
-        config.fetch("months_to_sync", 3)
       end
 
       def money_forward
@@ -62,8 +48,8 @@ module MFYNAB
 
       def session
         @_session ||= MoneyForward::Session.new(
-          username: config["moneyforward_username"],
-          password: config["moneyforward_password"],
+          username: config.credentials["moneyforward_username"],
+          password: config.credentials["moneyforward_password"],
           logger: logger,
         )
       end
@@ -75,15 +61,7 @@ module MFYNAB
       end
 
       def config
-        @_config ||= YAML
-          .load_file(config_file)
-          .values
-          .first
-          .merge(
-            "ynab_access_token" => ENV.fetch("YNAB_ACCESS_TOKEN"),
-            "moneyforward_username" => ENV.fetch("MONEYFORWARD_USERNAME"),
-            "moneyforward_password" => ENV.fetch("MONEYFORWARD_PASSWORD"),
-          )
+        @_config ||= Config.from_yaml(config_file, logger)
       end
 
       def logger