+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## (unreleased)
 
+- Credentials can now be stored in the config file.
+  (This change is backwards compatible.)
+- The configuration YAML file does not need a top-level key anymore.
+  (This change is backwards compatible.)
+
 ## 0.2.0 (2025-03-25)
 
 - Refresh accounts on Money Forward before syncing transactions to YNAB.